rename test data files to avoid redundancy
