+++ packages/devtools_app/test/cpu_profiler_tests/cpu_profile_model_test.dart
@@ -10,7 +10,7 @@ import 'package:devtools_test/devtools_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:vm_service/vm_service.dart';
 
-import '../test_data/cpu_profile_test_data.dart';
+import '../test_data/cpu_profile.dart';
 
 void main() {
   group('CpuProfileData', () {
