+++ packages/devtools_app/test/cpu_profiler_tests/cpu_profile_transformer_test.dart
@@ -7,7 +7,7 @@ import 'package:devtools_app/src/screens/profiler/cpu_profile_transformer.dart';
 import 'package:devtools_app/src/shared/profiler_utils.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_data/cpu_profile_test_data.dart';
+import '../test_data/cpu_profile.dart';
 
 void main() {
   group('CpuProfileTransformer', () {
