+++ packages/devtools_app/test/cpu_profiler_tests/cpu_profiler_controller_test.dart
@@ -12,7 +12,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:vm_service/vm_service.dart';
 
-import '../test_data/cpu_profile_test_data.dart';
+import '../test_data/cpu_profile.dart';
 
 void main() {
   final ServiceConnectionManager fakeServiceManager = FakeServiceManager(
