+++ packages/devtools_app/test/cpu_profiler_tests/cpu_profiler_test.dart
@@ -22,7 +22,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../test_data/cpu_profile_test_data.dart';
+import '../test_data/cpu_profile.dart';
 
 void main() {
   late CpuProfiler cpuProfiler;
