+++ packages/devtools_app/test/memory_tests/memory_screen_test.dart
@@ -22,8 +22,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:vm_service/vm_service.dart';
 
-import '../test_data/memory_test_allocation_data.dart';
-import '../test_data/memory_test_data.dart';
+import '../test_data/memory.dart';
+import '../test_data/memory_allocation.dart';
 
 void main() {
   late MemoryScreen screen;
