+++ packages/devtools_app/test/network_tests/network_model_test.dart
@@ -12,7 +12,7 @@ import 'package:devtools_test/devtools_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:vm_service/vm_service.dart';
 
-import '../test_data/network_test_data.dart';
+import '../test_data/network.dart';
 import '../test_utils/network_test_utils.dart';
 
 void main() {
