+++ packages/devtools_app/test/performance_tests/event_details_test.dart
@@ -16,7 +16,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../test_data/performance_test_data.dart';
+import '../test_data/performance.dart';
 
 void main() {
   const windowSize = Size(2000.0, 1000.0);
