+++ packages/devtools_app/test/performance_tests/performance_utils_test.dart
@@ -5,7 +5,7 @@
 import 'package:devtools_app/src/screens/performance/performance_utils.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_data/performance_test_data.dart';
+import '../test_data/performance.dart';
 
 void main() {
   group('PerformanceUtils', () {
