+++ packages/devtools_app/test/performance_tests/tabbed_performance_view_test.dart
@@ -25,7 +25,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:vm_service/vm_service.dart' as vm_service;
 
-import '../test_data/performance_test_data.dart';
+import '../test_data/performance.dart';
 
 void main() {
   FakeServiceManager fakeServiceManager;
