+++ packages/devtools_app/test/performance_tests/timeline_flame_chart_test.dart
@@ -24,7 +24,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:vm_service/vm_service.dart' as vm_service;
 
-import '../test_data/performance_test_data.dart';
+import '../test_data/performance.dart';
 
 void main() {
   FakeServiceManager fakeServiceManager;
