+++ packages/devtools_app/test/shared_tests/flame_chart_test.dart
@@ -17,8 +17,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_data/cpu_profile_test_data.dart';
-import '../test_data/performance_test_data.dart';
+import '../test_data/cpu_profile.dart';
+import '../test_data/performance.dart';
 
 void main() {
   const defaultZoom = 1.0;
