+++ packages/devtools_app/test/test_data/app_size/precompiler_trace.dart
