+++ packages/devtools_app/test/test_data/app_size/sizes.dart
